+++ dev/com.ibm.ws.security.javaeesec_fat/test-applications/JavaEESec.war/src/web/war/servlets/customform/get/forward/MultipleISCustomFormServlet.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package web.war.servlets.customform.forward;
+package web.war.servlets.customform.get.forward;
 import web.jar.base.FlexibleBaseServlet;
 import javax.security.enterprise.authentication.mechanism.http.CustomFormAuthenticationMechanismDefinition;
 import javax.security.enterprise.authentication.mechanism.http.LoginToContinue;
